Create Task 1: Numbergame
This is a java code which selects a random number between 1 to 100 and then ask the user to guess which number has been chosen if the input is high the  code will tell the user that the number is high and if the number is low it will tell the user that the number is low. Once the right number has been selected the code exits and gives the user the points gained according to rounds played.